Added support for WaterDog proxy
This commit offers great support for WaterDog proxy. How the support works:
You'll be noticed with the "kick early" option. This option will allow server owners to decide whether or not they want the kicked before locating server screen. They can turn it on for waterdog compatibility - Basically, this support will only occur if you have a proxy, and you put it onto the other servers you can connect to using /server in-game. This message will also occur if you're not yet whitelisted or added to maintenance mode.

Along with this, we've also pushed plugin version to 1.3.0.

Added new function: isKickedEarly() and setIsKickedEarly().

More availability and descriptions are in the configuration file. Very descriptive. And.. Also reorganized everything in the configuration file to ensure they know what's to come. And yes, you will need to backup your config.yml file, and add the updated configuration file. And.. That's it. No seriously, you're done reading!